requirements.txt
amqp==2.6.1
asgiref==3.3.4
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.7
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-jsonfield==1.4.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
idna==2.10
importlib-metadata==4.5.0
kombu==4.6.11
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
prompt-toolkit==3.0.18
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
redis-server==6.0.9
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.1
